refactor : delete 모달 중첩 오류 해결
